% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_disease_stage_duration.R
\name{he_add_disease_stage_duration}
\alias{he_add_disease_stage_duration}
\title{Add new row to disease stage duration matrix}
\usage{
he_add_disease_stage_duration(
  duration_matrix,
  disease_stage_distribution,
  num_animals_to_distribute
)
}
\arguments{
\item{duration_matrix}{matrix of days animals will spend in the specified
disease stage}

\item{disease_stage_distribution}{probability distribution of an animal
spending a given number of days in the specified disease stage}

\item{num_animals_to_distribute}{number of animals to distribute into time
spent in the specified disease stage}
}
\description{
Add new row to disease stage duration matrix
}
