% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_infected_net_pen.R
\name{he_add_infected_net_pen}
\alias{he_add_infected_net_pen}
\title{Add infected net pen to infected net pen tracking}
\usage{
he_add_infected_net_pen(
  infected_net_pen_info,
  net_pen_info,
  new_infected_net_pens,
  n_infected_animals_by_stage,
  infection_origin,
  simulation_day
)
}
\arguments{
\item{infected_net_pen_info}{data frame of information on infected net_pens}

\item{net_pen_info}{data frame of information on net_pens}

\item{new_infected_net_pens}{list of newly infected net pen ids}

\item{n_infected_animals_by_stage}{vector of numeric values indicating the
number of infected animals in each stage of infection (latent, clinical,
and subclinical)}

\item{infection_origin}{source of the infection can be "index",
"between-net_pen", or "between-farm"}

\item{simulation_day}{day of simulation that infection is occurring}
}
\value{
data frame of information on infected net_pens, appended with rows to
for new infections
}
\description{
Add infected net pen to infected net pen tracking
}
