% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_infection_stage_duration.R
\name{he_add_infection_stage_duration}
\alias{he_add_infection_stage_duration}
\title{Add new row to infection stage duration matrix}
\usage{
he_add_infection_stage_duration(
  duration_matrix,
  infection_stage_distribution,
  n_animals_to_distribute
)
}
\arguments{
\item{duration_matrix}{matrix of days animals will spend in the specified
infection stage}

\item{infection_stage_distribution}{probability distribution of an animal
spending a given number of days in the specified infection stage}

\item{n_animals_to_distribute}{number of animals to distribute into time
spent in the specified infection stage}
}
\value{
infection stage matrix with an additional row including with animals
distributed into durations spent in the stage
}
\description{
Add new row to infection stage duration matrix
}
