% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_infections_to_inf_farm_info.R
\name{he_add_infections_to_inf_farm_info}
\alias{he_add_infections_to_inf_farm_info}
\title{Add new infections to farm info matrix}
\usage{
he_add_infections_to_inf_farm_info(
  inf_farm_info,
  species_info,
  new_inf_farm_ids,
  num_inf_animals
)
}
\arguments{
\item{inf_farm_info}{matrix of information on infected farms}

\item{species_info}{matrix of information on species}

\item{new_inf_farm_ids}{list of newly infected farm ids}

\item{num_inf_animals}{number of animals infected in the infection process}
}
\value{
matrix of information on infected farms, appended with rows to
represent new infections
}
\description{
Add new infections to farm info matrix
}
