% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_apply_between_farm_netpen_infection.R
\name{he_apply_between_farm_netpen_infection}
\alias{he_apply_between_farm_netpen_infection}
\title{Infect new netpens}
\usage{
he_apply_between_farm_netpen_infection(
  farm_info,
  inf_farm_info,
  species_info,
  newly_infected_farm_ids,
  simulation_day,
  label
)
}
\arguments{
\item{farm_info}{matrix of farm information}

\item{inf_farm_info}{matrix of information on infected farms}

\item{species_info}{matrix of species information}

\item{newly_infected_farm_ids}{farm identifiers which should have newly
infected netpens}

\item{simulation_day}{day timestep in the simulation}

\item{label}{? contact method potentially?}
}
\value{
?
}
\description{
Infect new netpens
}
