% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_apply_infection_stage_distribution.R
\name{he_apply_infection_stage_distribution}
\alias{he_apply_infection_stage_distribution}
\title{Applies distribution to determine infection stage of newly infected animals}
\usage{
he_apply_infection_stage_distribution(new_inf_rows, species_info, stage)
}
\arguments{
\item{new_inf_rows}{matrix with rows representing newly infected netpens}

\item{species_info}{matrix of information on species}

\item{stage}{infection stage the newly infected animals are entering}
}
\value{
?
}
\description{
Applies distribution to determine infection stage of newly infected animals
}
