% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_between_farm_inf_prob_vec.R
\name{he_calculate_between_farm_inf_prob_vec}
\alias{he_calculate_between_farm_inf_prob_vec}
\title{Converts a between-farm infection probability matrix into an infection
probability vector}
\usage{
he_calculate_between_farm_inf_prob_vec(prob_matrix)
}
\arguments{
\item{prob_matrix}{symmetrical matrix of infection probabilities}
}
\value{
a vector of infection probabilities
}
\description{
Converts a between-farm infection probability matrix into an infection
probability vector
}
