% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_between_net_pen_infection_prob.R
\name{he_calculate_between_net_pen_infection_prob}
\alias{he_calculate_between_net_pen_infection_prob}
\title{Calculate probability of infection transmission between net pens within a farm}
\usage{
he_calculate_between_net_pen_infection_prob(
  net_pen_to_net_pen,
  infected_net_pen_info
)
}
\arguments{
\item{net_pen_to_net_pen}{daily probability for infection between net pens
transmission within a farm}

\item{infected_net_pen_info}{data frame of infected net pen information}
}
\value{
probability of transmission of infection between net pens
}
\description{
Calculate probability of infection transmission between net pens within a farm
}
