% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_between_netpen_infection_prob.R
\name{he_calculate_between_netpen_infection_prob}
\alias{he_calculate_between_netpen_infection_prob}
\title{Calculate probability of infection transmission between netpens within a farm}
\usage{
he_calculate_between_netpen_infection_prob(
  netpen_to_netpen,
  infected_netpen_info
)
}
\arguments{
\item{netpen_to_netpen}{daily probability for between-netpen infection
transmission within a farm}

\item{infected_netpen_info}{data frame of infected netpen information}
}
\value{
probability of between-netpen infection transmission
}
\description{
Calculate probability of infection transmission between netpens within a farm
}
