% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_inf_prob_matrix.R
\name{he_calculate_inf_prob_matrix}
\alias{he_calculate_inf_prob_matrix}
\title{Calculates a symmetrical matrix of farm infection probabilities}
\usage{
he_calculate_inf_prob_matrix(
  connectivity_matrix,
  farm_ids,
  farm_to_farm,
  vaccine_efficacy,
  farm_active,
  farm_susceptibility,
  farm_infectiousness,
  connectivity_matrix_type = "distance"
)
}
\arguments{
\item{connectivity_matrix}{connectivity matrix in the form of a seaway
distance matrix (in kilometers)or a hydroconnectivity matrix}

\item{farm_ids}{farm identifiers for which to calculate infection probabilities}

\item{farm_to_farm}{scaling parameter for between-farm infection transmission}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}

\item{farm_active}{logical indicating whether or not the farm is active}

\item{farm_susceptibility}{logical indicating whether or not the farm is susceptible}

\item{farm_infectiousness}{logical indicating whether or not the farm is infectious}

\item{connectivity_matrix_type}{type of connectivity matrix provided, either
distance matrix or hydroconnectivity matrix}
}
\value{
a symmetrical(?) matrix of farm infection probabilities
}
\description{
Calculates a symmetrical matrix of farm infection probabilities
}
