% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_inf_prob_vec.R
\name{he_calculate_inf_prob_vec}
\alias{he_calculate_inf_prob_vec}
\title{Converts an infection probability matrix into an infection probability vector}
\usage{
he_calculate_inf_prob_vec(prob_matrix)
}
\arguments{
\item{prob_matrix}{symmetrical matrix of infection probabilities}
}
\value{
a vector of infection probabilities
}
\description{
Converts an infection probability matrix into an infection probability vector
}
