% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/he_calculate_net_change_in_disease_stage_count.R
\name{he_calculate_net_change_in_disease_stage_count}
\alias{he_calculate_net_change_in_disease_stage_count}
\title{Calculate net change in number of animals in each disease stage}
\usage{
he_calculate_net_change_in_disease_stage_count(
  disease_stage_duration_matrices,
  n_newly_infected,
  clinically_infected_prop
)
}
\arguments{
\item{disease_stage_duration_matrices}{list of matrices containing the
disease stage duration for each net pen}

\item{n_newly_infected}{the number of newly infected animals for each
net pen}

\item{clinically_infected_prop}{the proportion of animals which will enter
the clinical stage upon infection, which is the same as 1 minus the
proportion of animals which will enter the subclinical stage upon
infection}
}
\value{
net change in the number of animals in each disease stage
}
\description{
Calculate net change in number of animals in each disease stage
}
