% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/he_calculate_reed_frost_intrafarm_transmission.R
\name{he_calculate_reed_frost_intrafarm_transmission}
\alias{he_calculate_reed_frost_intrafarm_transmission}
\title{Calculate Reed-Frost Intrafarm Transmission}
\usage{
he_calculate_reed_frost_intrafarm_transmission(z, nn, pp)
}
\arguments{
\item{z}{?}

\item{nn}{?}

\item{pp}{?}
}
\value{
?
}
\description{
Calculate Reed-Frost Intrafarm Transmission
}
