% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/he_calculate_subclinical_clinical_infection_split.R
\name{he_calculate_subclinical_clinical_infection_split}
\alias{he_calculate_subclinical_clinical_infection_split}
\title{Calculate number of newly subclinically and clinically infected animals}
\usage{
he_calculate_subclinical_clinical_infection_split(
  n_transitioning,
  clinically_infected_prop
)
}
\arguments{
\item{n_transitioning}{number of animals transitioning out of the latent stage}

\item{clinically_infected_prop}{proportion of animals which should become
clinically infected}
}
\value{
matrix with two columns, the first representing the number of newly
subclinically infected animals and the second representing the number of
newly clinically infected animals
}
\description{
Calculate number of newly subclinically and clinically infected animals
}
