% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_within_net_pen_infection_prob.R
\name{he_calculate_within_net_pen_infection_prob}
\alias{he_calculate_within_net_pen_infection_prob}
\title{Calculate probability of infection transmission within a net pen}
\usage{
he_calculate_within_net_pen_infection_prob(infected_net_pen_info)
}
\arguments{
\item{infected_net_pen_info}{data frame of infected net pen information}
}
\value{
probability of transmission of infection between net pens
}
\description{
Calculate probability of infection transmission within a net pen
}
