% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_within_netpen_infection_prob.R
\name{he_calculate_within_netpen_infection_prob}
\alias{he_calculate_within_netpen_infection_prob}
\title{Calculate probability of infection transmission within a netpen}
\usage{
he_calculate_within_netpen_infection_prob(infected_netpen_info)
}
\arguments{
\item{infected_netpen_info}{data frame of infected netpen information}
}
\value{
probability of within-netpen infection transmission
}
\description{
Calculate probability of infection transmission within a netpen
}
