% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_check_if_infection_is_resolved.R
\name{he_check_if_infection_is_resolved}
\alias{he_check_if_infection_is_resolved}
\title{Check if infection is resolved}
\usage{
he_check_if_infection_is_resolved(infected_net_pen_info)
}
\arguments{
\item{infected_net_pen_info}{data frame containing infected net pen info}
}
\value{
boolean value indicating whether the infection has run its course,
i.e. all fish have progressed to the immune stage (includes dead)
}
\description{
Check if infection is resolved
}
