% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_create_simulation_env.R
\name{he_create_simulation_env}
\alias{he_create_simulation_env}
\title{Create hydroepix simulation environment}
\usage{
he_create_simulation_env(model_env)
}
\arguments{
\item{model_env}{environment to be used as the environment parent, which
should be the model environment in the hydroepix context}
}
\value{
environment with provided argument environment as a parent
}
\description{
Create hydroepix simulation environment
}
