% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_depop_control_params.R
\name{he_define_depop_control_params}
\alias{he_define_depop_control_params}
\title{Define simulation variables related to depopulation control}
\usage{
he_define_depop_control_params(
  environment,
  depop_capacity = 20000,
  species_to_depop = c(1:18),
  depop_farm_if_netpen_infected = FALSE
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and
managed}

\item{depop_capacity}{integer value indicating the number of individuals
which can be depopulated per day}

\item{species_to_depop}{vector of indices to indicate which species can be depopulated??}

\item{depop_farm_if_netpen_infected}{boolean indicating whether the entire farm
should be depopulated if an infection is detected in at least one netpen}
}
\value{
NA
}
\description{
Define simulation variables related to depopulation control
}
