% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_depopulation_control_params.R
\name{he_define_depopulation_control_params}
\alias{he_define_depopulation_control_params}
\title{Define simulation variables related to depopulation control}
\usage{
he_define_depopulation_control_params(
  environment,
  culling_capacity = c(20000),
  species_to_cull = c(1:18),
  cull_farm_if_cage_infected = FALSE
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and
managed}

\item{culling_capacity}{vector of integer values indicating the number of
individuals which can be culled??}

\item{species_to_cull}{vector of indices to indicate which species can be culled??}

\item{cull_farm_if_cage_infected}{boolean indicating whether the entire farm
should be culled if an infection is detected in a cage}
}
\value{
NA
}
\description{
Define simulation variables related to depopulation control
}
