% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_simulation_control_params.R
\name{he_define_simulation_control_params}
\alias{he_define_simulation_control_params}
\title{Define simulation variables related to simulation control}
\usage{
he_define_simulation_control_params(
  model_env,
  n_simulations = 10,
  max_outbreak_length = 365,
  random_seed = -10,
  ignore_disease_status_input = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{model_env}{environment in which simulation variables are stored and
managed, this will be modified by the function}

\item{n_simulations}{number of simulations (i.e., individual simulated epidemics) to be run}

\item{max_outbreak_length}{maximum length of a disease outbreak}

\item{random_seed}{random seed to use for the simulation}

\item{ignore_disease_status_input}{boolean indicating whether or not to
ignore disease status input?}

\item{verbose}{boolean indicating whether verbose output should be generated}
}
\value{
NA
}
\description{
Define simulation variables related to simulation control
}
