% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_spread_control_params.R
\name{he_define_spread_control_params}
\alias{he_define_spread_control_params}
\title{Define simulation variables related to depopulation control}
\usage{
he_define_spread_control_params(
  environment,
  new_infection_functions = c("BCSpread()", "DBinf()"),
  t_stochastic = FALSE,
  rf_stochastic = FALSE,
  index_farm_function = "selectIndexFarm",
  index_farm_select = list(farm_type = 1:18),
  index_direct = FALSE,
  first_det_para_1 = 0.00255,
  first_det_para_2 = 0.00255,
  case_fatality = 0.89,
  days_dead_infectious = 2,
  farm_to_farm = 0.42,
  cage_to_cage = 0.052,
  vaccine_efficacy = 0
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and managed}

\item{new_infection_functions}{functions which define infection spread}

\item{t_stochastic}{?}

\item{rf_stochastic}{?}

\item{index_farm_function}{function which defines how the index farm is selected}

\item{index_farm_select}{arguments provided to the index_farm_function}

\item{index_direct}{?}

\item{first_det_para_1}{?}

\item{first_det_para_2}{?}

\item{case_fatality}{proportion of the infected population that will die}

\item{days_dead_infectious}{number of days dead animals remain infectious}

\item{farm_to_farm}{rate of infection between farms}

\item{cage_to_cage}{rate of infection between cages}

\item{vaccine_efficacy}{proportion of the vaccinated population that will survive if infected?}
}
\value{
NA
}
\description{
Define simulation variables related to depopulation control
}
