% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_surveillance_control_params.R
\name{he_define_surveillance_control_params}
\alias{he_define_surveillance_control_params}
\title{Define simulation variables related to depopulation control}
\usage{
he_define_surveillance_control_params(
  environment,
  control_functions = c("SurvZone(label='SZ')"),
  prob_diagnosis = 1,
  mortality_increase_1 = 2,
  mortality_increase_2 = 2,
  time_between_visits_within_zone = 4,
  protection_zone_duration = 50,
  past_days_for_dead_animal_surveillance = 7,
  daily_farm_survey_limit = 2,
  output_surveillance_details = FALSE
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and managed}

\item{control_functions}{functions which define infection control methods}

\item{prob_diagnosis}{probability of diagnosis given an infected individual is surveyed}

\item{mortality_increase_1}{?}

\item{mortality_increase_2}{?}

\item{time_between_visits_within_zone}{length of time between visits within a surveillance zone}

\item{protection_zone_duration}{length of time that an implemented protection zone lasts}

\item{past_days_for_dead_animal_surveillance}{?}

\item{daily_farm_survey_limit}{maximum number of farms that can be surveyed in a day}

\item{output_surveillance_details}{logical indicating whether to output detailed surveillance information}
}
\value{
NA
}
\description{
Define simulation variables related to depopulation control
}
