% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_generate_seaway_dist_prob.R
\name{he_generate_seaway_dist_prob}
\alias{he_generate_seaway_dist_prob}
\title{Calculates probability of infection given a set of seaway distances}
\usage{
he_generate_seaway_dist_prob(
  farm_id,
  dist_matrix,
  farm_to_farm,
  vaccine_efficacy
)
}
\arguments{
\item{farm_id}{farm identifiers to apply the calculation to}

\item{dist_matrix}{symmetrical matrix of seaway distances in kilometers}

\item{farm_to_farm}{scaling parameter for between-farm infection transmission}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}
}
\value{
scalar or vector of distance-scaled infection probabilities
}
\description{
Calculates probability of infection given a set of seaway distances
}
