% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/he_identify_netpens_for_between_farm_infection.R
\name{he_identify_netpens_for_between_farm_infection}
\alias{he_identify_netpens_for_between_farm_infection}
\title{Generates the subset of netpens which can be infected from a selection}
\usage{
he_identify_netpens_for_between_farm_infection(farm_info, inf_farm_id)
}
\arguments{
\item{farm_info}{matrix of farm information}

\item{inf_farm_id}{identifier of the farm which is to be infected}
}
\value{
vector of netpen identifiers for the given farm, where those netpens
can be validly infected
}
\description{
Generates the subset of netpens which can be infected from a selection
}
