% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_infect_netpens.R
\name{he_infect_netpens}
\alias{he_infect_netpens}
\title{Infect new netpens}
\usage{
he_infect_netpens(
  farm_info,
  inf_farm_info,
  species_info,
  newly_infected_farm_ids,
  g_time,
  label
)
}
\arguments{
\item{farm_info}{matrix of farm information}

\item{inf_farm_info}{matrix of information on infected farms}

\item{species_info}{matrix of species information}

\item{newly_infected_farm_ids}{farm identifiers which should have newly
infected netpens}

\item{g_time}{simulation timestep?}

\item{label}{? contact method potentially?}
}
\value{
?
}
\description{
Infect new netpens
}
