% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_farm_info.R
\name{he_initialize_farm_info}
\alias{he_initialize_farm_info}
\title{Initialize farm information from input files and parameters}
\usage{
he_initialize_farm_info(farm_info, species_info)
}
\arguments{
\item{farm_info}{data frame storing information on farms}

\item{species_info}{data frame storing information on species}
}
\value{
data frame of farm information initialized with necessary columns
and necessary species information for the entire model run
}
\description{
Initialize farm information from input files and parameters
}
