% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_inf_netpen_output.R
\name{he_initialize_inf_netpen_output}
\alias{he_initialize_inf_netpen_output}
\title{Initialize infected netpen output}
\usage{
he_initialize_inf_netpen_output(
  environment,
  filepath,
  inf_netpen_output_file_name = "infected_netpens.txt"
)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the infected netpen output file}

\item{inf_netpen_output_file_name}{name to use for creating and writing the
infected netpen output file}
}
\value{
NA
}
\description{
Initialize infected netpen output
}
