% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infected_cage_output.R
\name{he_initialize_infected_cage_output}
\alias{he_initialize_infected_cage_output}
\title{Initialize infected cage output}
\usage{
he_initialize_infected_cage_output(environment, filename)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the infected cage output file}
}
\value{
NA
}
\description{
Initialize infected cage output
}
