% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infected_fish_single.R
\name{he_initialize_infected_fish_single}
\alias{he_initialize_infected_fish_single}
\title{Initialize infection as a single random fish}
\usage{
he_initialize_infected_fish_single(
  contact_farms,
  prob_infection,
  susceptibility
)
}
\arguments{
\item{contact_farms}{which farms have had contact with the vector of infection?}

\item{prob_infection}{probability of infection of ...?}

\item{susceptibility}{susceptibility? to infection? logical yes/no?}
}
\value{
?
}
\description{
Initialize infection as a single random fish
}
