% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_inf_fish_stochastic.R
\name{he_initialize_infected_fish_stochastic}
\alias{he_initialize_infected_fish_stochastic}
\title{Initialize infection as a randomized number of random fish}
\usage{
he_initialize_infected_fish_stochastic(
  contact_farms,
  prob_infection,
  susceptibility
)
}
\arguments{
\item{contact_farms}{which farms have had contact with the vector of infection?}

\item{prob_infection}{probability of infection of ...?}

\item{susceptibility}{susceptibility? to infection? logical yes/no?}
}
\value{
NA
}
\description{
Initialize infection as a randomized number of random fish
}
