% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infected_net_pen_info.R
\name{he_initialize_infected_net_pen_info}
\alias{he_initialize_infected_net_pen_info}
\title{Initialize data frame to store and track infected net pen information}
\usage{
he_initialize_infected_net_pen_info(output_dir, output_file_name)
}
\arguments{
\item{output_dir}{file path where the infected net pen information output should
be stored}

\item{output_file_name}{name of the file in which the infected net pen information
should be stored}
}
\value{
empty data frame with appropriate column headers and types to store
infected net pen information
}
\description{
Initialize data frame to store and track infected net pen information
}
