% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infected_netpen_output.R
\name{he_initialize_infected_netpen_output}
\alias{he_initialize_infected_netpen_output}
\title{Initialize infected netpen output}
\usage{
he_initialize_infected_netpen_output(environment, filename)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the infected netpen output file}
}
\value{
NA
}
\description{
Initialize infected netpen output
}
