% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_internal_model_vars.R
\name{he_initialize_internal_model_vars}
\alias{he_initialize_internal_model_vars}
\title{Initialize internal simulation variables}
\usage{
he_initialize_internal_model_vars(environment)
}
\arguments{
\item{environment}{simulation environment}
}
\value{
NA
}
\description{
Initialize internal simulation variables
}
