% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_model_env.R
\name{he_initialize_model_env}
\alias{he_initialize_model_env}
\title{Initialize model variables}
\usage{
he_initialize_model_env(
  environment,
  farm_info_filepath,
  species_info_filepath,
  dist_mat_filepath,
  output_filepath = "/output",
  ignore_status = TRUE
)
}
\arguments{
\item{environment}{environment in which to store model level variables}

\item{farm_info_filepath}{filepath to farm info file}

\item{species_info_filepath}{filepath to species info file}

\item{dist_mat_filepath}{filepath to distance matrix file}

\item{output_filepath}{filepath to the directory in which to write output files}

\item{ignore_status}{logical value indicating whether to ignore infection
status values provided in input files or not}

\item{detailed_survey_output}{logical value indicating whether to generate
detailed surveillance output}

\item{detailed_depop_output}{logical value indicating whether to generate
detailed depopulation output}

\item{detailed_preemptive_depop_output}{logical value indicating whether to
generate detailed preemptive depopulation output}
}
\value{
NA
}
\description{
Initialize model variables
}
