% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_net_pen_info.R
\name{he_initialize_net_pen_info}
\alias{he_initialize_net_pen_info}
\title{Initialize net pen information from input files and parameters}
\usage{
he_initialize_net_pen_info(net_pen_info, species_info)
}
\arguments{
\item{net_pen_info}{data frame storing information on net pens}

\item{species_info}{data frame storing information on species}
}
\value{
data frame of net pen information initialized with necessary columns
and necessary species information for the entire model run
}
\description{
Initialize net pen information from input files and parameters
}
