% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_netpen_info.R
\name{he_initialize_netpen_info}
\alias{he_initialize_netpen_info}
\title{Initialize netpen information from input files and parameters}
\usage{
he_initialize_netpen_info(netpen_info, species_info)
}
\arguments{
\item{netpen_info}{data frame storing information on netpens}

\item{species_info}{data frame storing information on species}
}
\value{
data frame of netpen information initialized with necessary columns
and necessary species information for the entire model run
}
\description{
Initialize netpen information from input files and parameters
}
