% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_preemption_output.R
\name{he_initialize_preemption_output}
\alias{he_initialize_preemption_output}
\title{Initialize preemption output}
\usage{
he_initialize_preemption_output(environment, filepath)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the preemption output file}
}
\value{
NA
}
\description{
Initialize preemption output
}
