% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_preemptive_depop_output.R
\name{he_initialize_preemptive_depop_output}
\alias{he_initialize_preemptive_depop_output}
\title{Initialize preemptive depopulation output}
\usage{
he_initialize_preemptive_depop_output(environment, filepath)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the preemption depopulation output file}
}
\value{
NA
}
\description{
Initialize preemptive depopulation output
}
