% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_result_summary_output.R
\name{he_initialize_result_summary_output}
\alias{he_initialize_result_summary_output}
\title{Initialize result summary output}
\usage{
he_initialize_result_summary_output(environment, filepath)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the results summary output file}
}
\value{
NA
}
\description{
Initialize result summary output
}
