% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_survey_output.R
\name{he_initialize_survey_output}
\alias{he_initialize_survey_output}
\title{Initialize survey output}
\usage{
he_initialize_survey_output(environment, filepath)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the survey output file}
}
\value{
NA
}
\description{
Initialize survey output
}
