% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_parse_disease_stage_distributions.R
\name{he_parse_disease_stage_distributions}
\alias{he_parse_disease_stage_distributions}
\title{Parse Disease Stage Distributions}
\usage{
he_parse_disease_stage_distributions(species_info_row)
}
\arguments{
\item{species_info_row}{a row of species info data, including headers}
}
\value{
species info data frame with disease distribution column inputs parsed
from literal text to evaluated code expressions
}
\description{
Parse Disease Stage Distributions
}
