% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_parse_infection_stage_distributions.R
\name{he_parse_infection_stage_distributions}
\alias{he_parse_infection_stage_distributions}
\title{Parse infection stage distributions}
\usage{
he_parse_infection_stage_distributions(species_info_row)
}
\arguments{
\item{species_info_row}{a row of species info data, including headers}
}
\value{
species info data frame with infection distribution column inputs parsed
from literal text to evaluated code expressions
}
\description{
Parse infection stage distributions
}
