% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_plot_simulations.R
\name{he_plot_simulation}
\alias{he_plot_simulation}
\title{Plot the number of individuals in each stage over the simulation}
\usage{
he_plot_simulation(dat, pal = NULL)
}
\arguments{
\item{dat}{tbd}
}
\value{
tbd
}
\description{
Infection stages will have the same colour between plots, even if a stage(s)
is dropped.
}
