% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_plot_simulation_examples.R
\name{he_plot_simulation_examples}
\alias{he_plot_simulation_examples}
\title{Plot the infection stages of each net pen over up to 5 example simulations}
\usage{
he_plot_simulation_examples(
  sim_dat,
  simulation_nums = NULL,
  infection_stages = NULL
)
}
\arguments{
\item{sim_dat}{simulation data, expected in the format from using the he_read_simulation_output_data function}

\item{simulation_nums}{simulation number(s) of the simulation to plot}

\item{infection_stages}{stage(s) of infection to plot. Options are "susceptible", "latent", "subclinical", "clinical", "dead", and "recovered"}
}
\value{
ggplot object with multiple faceted plots
}
\description{
Plot the infection stages of each net pen over up to 5 example simulations
}
