% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_connectivity_matrix_file.R
\name{he_read_connectivity_matrix_file}
\alias{he_read_connectivity_matrix_file}
\title{Read in a connectivity matrix file}
\usage{
he_read_connectivity_matrix_file(
  filepath,
  connectivity_matrix_type = "distance"
)
}
\arguments{
\item{filepath}{a filepath to read a csv file from}

\item{connectivity_matrix_type}{type of connectivity matrix provided, either
distance matrix or hydroconnectivity matrix}
}
\value{
data.frame of the csv file
}
\description{
Read in a connectivity matrix file
}
