% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_dist_mat_file.R
\name{he_read_dist_mat_file}
\alias{he_read_dist_mat_file}
\title{Read in a distance matrix file}
\usage{
he_read_dist_mat_file(filepath)
}
\arguments{
\item{filepath}{a filepath to read a csv file from}
}
\value{
data.frame of the csv file
}
\description{
Read in a distance matrix file
}
