% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_infected_net_pen_output_data.R
\name{he_read_infected_net_pen_output_data}
\alias{he_read_infected_net_pen_output_data}
\title{Read infected net pen output files}
\usage{
he_read_infected_net_pen_output_data(
  infected_net_pen_output_file_name,
  output_dir
)
}
\arguments{
\item{infected_net_pen_output_file_name}{name of the infected net pen output files, should be the same as the one defined by the user in the he_define_output_params function}

\item{output_dir}{directory containing the specified infected net pen output files, likely the same as the one defined by the user in the he_define_output_params function}
}
\value{
data.frame of up to 5 infected net pen output simulation files stacked together
}
\description{
Read infected net pen output files
}
