% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_netpen_info_file.R
\name{he_read_netpen_info_file}
\alias{he_read_netpen_info_file}
\title{Read in a netpen information file}
\usage{
he_read_netpen_info_file(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{a filepath to read a csv file from}

\item{verbose}{a logical value indicating whether to provide additional
output messages on function progress}
}
\value{
data.frame of the csv file
}
\description{
Read in a netpen information file
}
