% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_types_file.R
\name{he_read_species_info_file}
\alias{he_read_species_info_file}
\title{Read Species Info File}
\usage{
he_read_species_info_file(filepath)
}
\arguments{
\item{filepath}{to species info file}
}
\value{
data frame of species info
}
\description{
Read Species Info File
}
