% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_read_types_file.R
\name{he_read_types_file}
\alias{he_read_types_file}
\title{Read Types File}
\usage{
he_read_types_file(filepath)
}
\arguments{
\item{filepath}{to types file}
}
\value{
data frame of type info from types file
}
\description{
Read Types File
}
