% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_reset_simulation_env.R
\name{he_reset_simulation_env}
\alias{he_reset_simulation_env}
\title{Reset the simulation environment at the end of a simulation}
\usage{
he_reset_simulation_env(environment, num_farms, farm_info)
}
\arguments{
\item{environment}{simulation environment to reset}

\item{num_farms}{number of farms in the model run}

\item{farm_info}{matrix of farm information}
}
\value{
matrix of farm information, reset to simulation start values
}
\description{
Reset the simulation environment at the end of a simulation
}
