% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_retrieve_susceptible_net_pens.R
\name{he_retrieve_susceptible_net_pens}
\alias{he_retrieve_susceptible_net_pens}
\title{Simulate a day of infection transmission dynamics}
\usage{
he_retrieve_susceptible_net_pens(farm_id, net_pen_info, infected_net_pen_info)
}
\arguments{
\item{farm_id}{id of farm within which between-net-pen infection is to be simulated}

\item{net_pen_info}{data frame of information on net pens}

\item{infected_net_pen_info}{data frame of information on infected net pens}
}
\value{
number of net pens which are susceptible to infection in the specified farm
}
\description{
Simulate a day of infection transmission dynamics
}
