% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_rpert.R
\name{he_rpert}
\alias{he_rpert}
\title{Generate n PERT beta random variates}
\usage{
he_rpert(n, min, mode, max)
}
\arguments{
\item{n}{number of PERT beta random variates to generate}

\item{min}{minimum value}

\item{mode}{mode value}

\item{max}{maximum value}
}
\value{
vector of PERT beta random variates
}
\description{
Generate n PERT beta random variates
}
\examples{
he_rpert(10, 0.14, 0.4, 0.8)
}
