% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_rpoly2.R
\name{he_rpoly2}
\alias{he_rpoly2}
\title{Polynomial distribution sampling function}
\usage{
he_rpoly2(sample_n, probability)
}
\arguments{
\item{sample_n}{number of samples to take}

\item{probability}{numeric probability or vector of probabilities}
}
\value{
number of samples distributed into each bin by probability
}
\description{
Polynomial distribution sampling function
}
