% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_run_model.R
\name{he_run_model}
\alias{he_run_model}
\title{Run hydroepix model}
\usage{
he_run_model(model_env)
}
\arguments{
\item{model_env}{environment containing variables for all essential hydroepix
model parameters}
}
\description{
Run hydroepix model
}
