% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_select_index_farm.R
\name{he_select_index_farm}
\alias{he_select_index_farm}
\title{Select Index Farm}
\usage{
he_select_index_farm(farm_info, farm_id = NULL, num_inf_netpens = 1)
}
\arguments{
\item{farm_info}{data frame of farm information}

\item{farm_id}{identifier for farm to infect}

\item{num_inf_netpens}{}
}
\value{
netpen identifier(s) of initially infected netpens
}
\description{
Select Index Farm
}
