% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_select_index_net_pens.R
\name{he_select_index_net_pens}
\alias{he_select_index_net_pens}
\title{Select Index Net Pens}
\usage{
he_select_index_net_pens(
  net_pen_info,
  net_pen_ids = NULL,
  farm_id = NULL,
  n_infected_net_pens = 1
)
}
\arguments{
\item{net_pen_info}{data frame of net pen information}

\item{net_pen_ids}{identifiers of specific net pens to be infected}

\item{farm_id}{identifier for farm to be infected first, if no net pen
identifiers are supplied}

\item{n_infected_net_pens}{number of randomly selected net pens to be infected in
the specified farm, if no net pen identifiers are supplied}
}
\value{
net pen identifier(s) of initially infected net pens
}
\description{
Select Index Net Pens
}
