% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_select_index_netpens.R
\name{he_select_index_netpens}
\alias{he_select_index_netpens}
\title{Select Index Netpens}
\usage{
he_select_index_netpens(
  farm_info,
  netpen_ids = NULL,
  farm_id = NULL,
  n_inf_netpens = 1
)
}
\arguments{
\item{farm_info}{data frame of farm information}

\item{netpen_ids}{identifiers of specific netpens to be infected}

\item{farm_id}{identifier for farm to be infected first, if no netpen
identifiers are supplied}

\item{n_inf_netpens}{number of randomly selected netpens to be infected in
the specified farm, if no netpen identifiers are supplied}
}
\value{
netpen identifier(s) of initially infected netpens
}
\description{
Select Index Netpens
}
