% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_set_random_seed.R
\name{he_set_random_seed}
\alias{he_set_random_seed}
\title{Set random seed}
\usage{
he_set_random_seed(seed, simulation_n)
}
\arguments{
\item{seed}{numeric value for the seed}

\item{simulation_n}{simulation number, based on the number of simulations
the model is set to run}
}
\value{
seed value, variable by simulation number if seed value provided is
negative, fixed otherwise
}
\description{
Set random seed
}
