% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_simulate_day.R
\name{he_simulate_day}
\alias{he_simulate_day}
\title{Simulate a day of infection transmission dynamics}
\usage{
he_simulate_day(
  inf_farm_info,
  simulation_env,
  simulation_day,
  species_info,
  verbose = FALSE
)
}
\arguments{
\item{inf_farm_info}{data frame of information on infected farms}

\item{simulation_env}{environment containing simulation variables}

\item{simulation_day}{day of simulation that infection is occurring}

\item{species_info}{matrix of species information}

\item{verbose}{a logical value indicating whether to provide additional output
messages on model running progress}
}
\description{
Simulate a day of infection transmission dynamics
}
