% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_update_disease_stage_counts.R
\name{he_update_disease_stage_counts}
\alias{he_update_disease_stage_counts}
\title{Update counts of animals transitioning from one disease stage into another}
\usage{
he_update_disease_stage_counts(
  disease_stage_counts,
  disease_stage_duration_matrices,
  num_newly_infected
)
}
\arguments{
\item{disease_stage_counts}{data frame containing the counts of animals in
each disease stage, where each disease stage is a column and each row is
a netpen}

\item{disease_stage_duration_matrices}{list of matrices containing the
disease stage duration for each netpen}

\item{num_newly_infected}{the number of newly infected animals for each
netpen}
}
\value{
modified disease stage count columns
}
\description{
Update counts of animals transitioning from one disease stage into another
}
