% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_update_infection_stage_counts.R
\name{he_update_infection_stage_counts}
\alias{he_update_infection_stage_counts}
\title{Update counts of animals transitioning from one infection stage into another}
\usage{
he_update_infection_stage_counts(
  infection_stage_counts,
  infection_stage_duration_matrices,
  n_newly_infected,
  clinically_infected_prop
)
}
\arguments{
\item{infection_stage_counts}{data frame containing the counts of animals in
each infection stage, where each infection stage is a column and each row is
a net pen}

\item{infection_stage_duration_matrices}{list of matrices containing the
infection stage duration for each net pen}

\item{n_newly_infected}{the number of newly infected animals for each
net pen}

\item{clinically_infected_prop}{the proportion of animals which will enter
the clinical stage upon infection, which is the same as 1 minus the
proportion of animals which will enter the subclinical stage upon
infection}
}
\value{
modified infection stage count columns
}
\description{
Update counts of animals transitioning from one infection stage into another
}
