% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_update_infection_stage_duration_matrix.R
\name{he_update_infection_stage_duration_matrix}
\alias{he_update_infection_stage_duration_matrix}
\title{Update infection duration stage matrices with newly distributed animals}
\usage{
he_update_infection_stage_duration_matrix(
  infection_stage_matrix,
  infection_stage_distribution,
  n_animals_to_distribute
)
}
\arguments{
\item{infection_stage_matrix}{matrix representing a infection stage, recording
the number of animals that will spend a given number of days in that
infection stage}

\item{infection_stage_distribution}{data frame of species information}

\item{n_animals_to_distribute}{number of animals which are newly entering
the given infection stage}
}
\value{
updated infection stage duration matrix with newly distributed animals
}
\description{
Update infection duration stage matrices with newly distributed animals
}
