% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_inf_netpen_output.R
\name{he_write_inf_netpen_output}
\alias{he_write_inf_netpen_output}
\title{Write infected netpen output to file}
\usage{
he_write_inf_netpen_output(
  inf_farm_info,
  inf_netpen_output_file_name,
  output_dir
)
}
\arguments{
\item{inf_farm_info}{data frame containing infected netpen data, initialized
and updated over the course of a simulation run}

\item{inf_netpen_output_file_name}{string to name the infected netpen output
file}

\item{output_dir}{string to indicate the directory in which the output file
should be written}
}
\value{
NA
}
\description{
Write infected netpen output to file
}
