% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_infected_net_pen_output.R
\name{he_write_infected_net_pen_output}
\alias{he_write_infected_net_pen_output}
\title{Write infected net pen output to file}
\usage{
he_write_infected_net_pen_output(
  infected_net_pen_info,
  infected_net_pen_output_file_name,
  output_dir
)
}
\arguments{
\item{infected_net_pen_info}{data frame containing infected net pen data, initialized
and updated over the course of a simulation run}

\item{infected_net_pen_output_file_name}{string to name the infected net pen output
file}

\item{output_dir}{string to indicate the directory in which the output file
should be written}
}
\value{
NA
}
\description{
Write infected net pen output to file
}
