% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_output_columns.R
\name{he_write_output_columns}
\alias{he_write_output_columns}
\title{Write names of columns to output file}
\usage{
he_write_output_columns(output_column_names, output_dir, output_file_name)
}
\arguments{
\item{output_column_names}{names of columns to write to file}

\item{output_dir}{directory of file to be written to}

\item{output_file_name}{name of file to be written to}
}
\value{
NA
}
\description{
Write names of columns to output file
}
