% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_preemptive_depop_output.R
\name{he_write_preemptive_depop_output}
\alias{he_write_preemptive_depop_output}
\title{Write pre-emptive depopulation output to file}
\usage{
he_write_preemptive_depop_output(
  preemptive_depop_matrix,
  preemptive_depop_file_name = "preemptive_depop.txt",
  output_dir = "",
  append = FALSE
)
}
\arguments{
\item{preemptive_depop_matrix}{matrix containing depopulation variables, initialized
and updated over the course of a simulation run}

\item{preemptive_depop_file_name}{string to name the depopulation output file}

\item{output_dir}{string to indicate the directory in which the output file
should be written}

\item{append}{logical indicating whether to overwrite an existing file or
append data}
}
\value{
NA
}
\description{
Write pre-emptive depopulation output to file
}
