% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_survey_output.R
\name{he_write_survey_output}
\alias{he_write_survey_output}
\title{Write survey output to file}
\usage{
he_write_survey_output(
  surveyed_output_matrix,
  survey_output_file_name = "surveyed_farms.txt",
  output_dir = "",
  append = FALSE
)
}
\arguments{
\item{survey_output_file_name}{string to name the depopulation output file}

\item{output_dir}{string to indicate the directory in which the output file
should be written}

\item{append}{logical indicating whether to overwrite an existing file or
append data}

\item{survey_matrix}{matrix containing surveyed farm variables, initialized
and updated over the course of a simulation run}
}
\value{
NA
}
\description{
Write survey output to file
}
