% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_farm_info_file.R
\name{read_farm_info_file}
\alias{read_farm_info_file}
\title{Read in a farm information file}
\usage{
read_farm_info_file(filepath)
}
\arguments{
\item{filepath}{a filepath to read a csv file from}
}
\value{
data.frame of the csv file
}
\description{
Read in a farm information file
}
\examples{
test1 <- c(1:5, "6,7", "8,9,10")
tf <- tempfile()
writeLines(test1, tf)
read_farm_info_file(tf)
}
