% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_types_file.R
\name{read_types_file}
\alias{read_types_file}
\title{Read Types File}
\usage{
read_types_file(filepath)
}
\arguments{
\item{filepath}{to types file}
}
\value{
data frame of type info from types file
}
\description{
Read Types File
}
\examples{
test_types_file <- c(
  1,
  "Fish1",
  "c(0.000001,0.000001,0.0001,0.01,0.1,0.25,0.339898,0.3)",
  "c(0.249995,0.50,0.25,0.000001,0.000001,0.000001,0.000001,0.000001)",
  "c(0.125,0.125,0.125,0.125,0.125,0.125,0.125,0.125)",
  "rpert(n,0.14,0.4,0.8)",
  1
)
tf <- tempfile()
writeLines(test_types_file, tf)
read_farm_info_file(tf)
}
