% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpert.R
\name{rpert}
\alias{rpert}
\title{Generate n PERT beta random variates}
\usage{
rpert(n, a, l, b)
}
\arguments{
\item{n}{number of PERT beta random variates to generate}

\item{a}{minimum value}

\item{l}{mode value}

\item{b}{maximum value}
}
\value{
vector of PERT beta random variates
}
\description{
Generate n PERT beta random variates
}
\examples{
rpert(10, 0.14, 0.4, 0.8)
}
