% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_infected_netpen.R
\name{he_add_infected_netpen}
\alias{he_add_infected_netpen}
\title{Add infected netpen to infected netpen tracking}
\usage{
he_add_infected_netpen(
  infected_netpen_info,
  netpen_info,
  new_infected_netpens,
  n_infected_animals_by_stage,
  infection_origin,
  simulation_day
)
}
\arguments{
\item{infected_netpen_info}{data frame of information on infected netpens}

\item{netpen_info}{data frame of information on netpens}

\item{new_infected_netpens}{list of newly infected netpen ids}

\item{n_infected_animals_by_stage}{vector of numeric values indicating the
number of infected animals in each stage of infection (latent, clinical,
and subclinical)}

\item{infection_origin}{source of the infection was can be "index",
between-netpen", or "between-farm"}

\item{simulation_day}{day of simulation that infection is occurring}
}
\value{
data frame of information on infected netpens, appended with rows to
for new infections
}
\description{
Add infected netpen to infected netpen tracking
}
