% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_add_infectd_netpen.R
\name{he_add_infected_netpen}
\alias{he_add_infected_netpen}
\title{Add infected netpen to the infected farm info data frame}
\usage{
he_add_infected_netpen(
  inf_farm_info,
  species_info,
  netpen_ids_to_infect,
  num_inf_animals_by_stage,
  type_of_contact
)
}
\arguments{
\item{inf_farm_info}{data frame of information on infected farms}

\item{species_info}{data frame of information on species}

\item{netpen_ids_to_infect}{list of newly infected netpen ids}

\item{num_inf_animals_by_stage}{vector of three numeric values indicating the
of infected animals in each stage of infection (latent, clinical, and
subclinical)}

\item{type_of_contact}{type of contact that resulted in the infection, can be
either direct (within/between netpens) or indirect (between farms)}
}
\value{
data frame of information on infected farms, appended with rows to
represent new infections
}
\description{
Add infected netpen to the infected farm info data frame
}
