% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_inf_prob_matrix.R
\name{he_calculate_inf_prob_matrix}
\alias{he_calculate_inf_prob_matrix}
\title{Calculates a symmetrical matrix of farm infection probabilities}
\usage{
he_calculate_inf_prob_matrix(
  dist_mat,
  farm_ids,
  farm_to_farm,
  vaccine_efficacy,
  farm_active,
  farm_susceptibility,
  farm_infectiousness,
  dist_mat_type = "seaway distance"
)
}
\arguments{
\item{dist_mat}{symmetrical matrix of distances between farms in kilometers?}

\item{farm_ids}{farm identifiers for which to calculate infection probabilities}

\item{farm_to_farm}{scaling parameter for between-farm infection transmission}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}

\item{farm_active}{logical indicating whether or not the farm is actively producing?}

\item{farm_susceptibility}{logical indicating whether or not the farm is susceptible}

\item{farm_infectiousness}{logical indicating whether or not the farm is infectious}

\item{dist_mat_type}{type of distance matrix provided (i.e. is this a distance
matrix or a matrix with using a different measure of connectivity)}
}
\value{
a symmetrical matrix of farm infection probabilities
}
\description{
Calculates a symmetrical matrix of farm infection probabilities
}
