% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/he_calculate_net_change_in_disease_stage_count.R
\name{he_calculate_net_change_in_disease_stage_count}
\alias{he_calculate_net_change_in_disease_stage_count}
\title{Calculate net change in number of animals in each disease stage}
\usage{
he_calculate_net_change_in_disease_stage_count(
  disease_stage_duration_matrices,
  n_newly_infected
)
}
\arguments{
\item{disease_stage_duration_matrices}{list of matrices containing the
disease stage duration for each netpen}

\item{n_newly_infected}{he number of newly infected animals for each
netpen}
}
\value{
net change in the number of animals in each disease stage
}
\description{
Calculate net change in number of animals in each disease stage
}
