% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_prevalence.R
\name{he_calculate_prevalence}
\alias{he_calculate_prevalence}
\title{Calculate the prevalence of infection}
\usage{
he_calculate_prevalence(n_infectious, n_not_immune)
}
\arguments{
\item{n_infectious}{number of animals which are infectious}

\item{n_not_immune}{number of animals which cannot be infected}
}
\value{
prevalence of infected animals within a net pen or a farm
}
\description{
Calculate the prevalence of infection
}
