% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_calculate_within_netpen_infection_prob.R
\name{he_calculate_within_netpen_infection_prob}
\alias{he_calculate_within_netpen_infection_prob}
\title{Calculate probability of infection transmission within a netpen}
\usage{
he_calculate_within_netpen_infection_prob(inf_farm_info, vaccine_efficacy)
}
\arguments{
\item{inf_farm_info}{data frame of infected farm information}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}
}
\value{
probability of within-netpen infection transmission
}
\description{
Calculate probability of infection transmission within a netpen
}
