% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_simulation_control_params.R
\name{he_define_simulation_control_params}
\alias{he_define_simulation_control_params}
\title{Define simulation variables related to simulation control}
\usage{
he_define_simulation_control_params(
  environment,
  n = 1,
  run_id = NULL,
  step_in_file = NULL,
  max_outbreak_length = 365,
  random_seed = -10,
  ignore_disease_status_input = TRUE,
  verbose = FALSE,
  summary_function = "HEsum"
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and
managed}

\item{n}{number of iterations?}

\item{run_id}{identifier for this run of the model}

\item{step_in_file}{filename for step-wise information?}

\item{max_outbreak_length}{maximum length of a disease outbreak}

\item{random_seed}{random seed to use for the simulation}

\item{ignore_disease_status_input}{boolean indicating whether or not to
ignore disease status input?}

\item{verbose}{boolean indicating whether verbose output should be generated}

\item{summary_function}{text indicating which summary function should be used
to summarize simulation outputs}
}
\value{
NA
}
\description{
Define simulation variables related to simulation control
}
