% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_spread_control_params.R
\name{he_define_spread_control_params}
\alias{he_define_spread_control_params}
\title{Define simulation variables related to spread control}
\usage{
he_define_spread_control_params(
  environment,
  random_num_initially_infected_fish = FALSE,
  intrafarm_disease_transmission_model = "binomial chain",
  index_farm_id = NULL,
  index_direct = TRUE,
  case_fatality_prop = 0.89,
  days_dead_infectious = 2,
  farm_to_farm = 0.42,
  netpen_to_netpen = 0.052,
  vaccine_efficacy = 0
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and
managed}

\item{random_num_initially_infected_fish}{logical indicating whether to
randomize the initial number of infected fish}

\item{intrafarm_disease_transmission_model}{defines which intrafarm disease
transmission model is used, options are "binomial chain" and "reed-frost"}

\item{index_farm_id}{identifier of the index farm}

\item{index_direct}{logical indicating whether the index farm gets infected
directly, as opposed to indirectly}

\item{case_fatality_prop}{proportion of the infected population that will die}

\item{days_dead_infectious}{number of days dead animals remain infectious}

\item{farm_to_farm}{scaling parameter for between-farm infection transmission}

\item{netpen_to_netpen}{daily probability for between-netpen infection
transmission within a farm}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}
}
\value{
NA
}
\description{
Define simulation variables related to spread control
}
