% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_define_surveillance_control_params.R
\name{he_define_surveillance_control_params}
\alias{he_define_surveillance_control_params}
\title{Define simulation variables related to surveillance control}
\usage{
he_define_surveillance_control_params(
  environment,
  control_functions = c("SurvZone(label='SZ')"),
  prob_diagnosis = 0.95,
  mort_increase_factor_for_first_investigation = 2,
  mort_increase_factor_for_subsequent_investigation = 2,
  time_between_survey_days = 4,
  max_duration_surveillance_zone_days = 50,
  past_days_for_dead_animal_surveillance = 7,
  daily_farm_survey_limit = 2,
  output_surveillance_details = FALSE
)
}
\arguments{
\item{environment}{environment in which simulation variables are stored and
managed}

\item{control_functions}{functions which define infection control methods}

\item{prob_diagnosis}{probability of diagnosis given an infected individual
is surveyed}

\item{mort_increase_factor_for_first_investigation}{factor of increase in
mortality required to trigger a first investigation}

\item{mort_increase_factor_for_subsequent_investigation}{factor of increase
in mortality required to trigger a subsequent investigation}

\item{time_between_survey_days}{length of time between visits within a
surveillance zone in days}

\item{max_duration_surveillance_zone_days}{length of time that an implemented protection
zone lasts in days}

\item{past_days_for_dead_animal_surveillance}{?}

\item{daily_farm_survey_limit}{maximum number of farms that can be surveyed
in a day in each bay}

\item{output_surveillance_details}{logical indicating whether to output
detailed surveillance information}
}
\value{
NA
}
\description{
Define simulation variables related to surveillance control
}
