% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_dist_based_infection.R
\name{he_dist_based_infection}
\alias{he_dist_based_infection}
\title{Determines infection spread between farms for a given day?}
\usage{
he_dist_based_infection(
  farm_info,
  sim_day,
  connectivity_matrix,
  farm_to_farm,
  vaccine_efficacy,
  connectivity_matrix_type = "distance",
  label = 1,
  t_start = 0,
  t_end = Inf
)
}
\arguments{
\item{farm_info}{matrix of farm and netpen information}

\item{sim_day}{day timestep in the simulation}

\item{connectivity_matrix}{connectivity matrix in the form of a seaway
distance matrix (in kilometers)or a hydroconnectivity matrix}

\item{farm_to_farm}{scaling parameter for between-farm infection transmission}

\item{vaccine_efficacy}{product of the manufacturer-reported vaccine efficacy
and the population coverage of the vaccine}

\item{connectivity_matrix_type}{type of connectivity matrix provided, either
distance matrix or hydroconnectivity matrix}

\item{label}{? seems to apply to infection mode somehow? (direct or indirect)?}

\item{t_start}{threshold of earliest possible infection spread}

\item{t_end}{threshold for latest possible infection spread}
}
\value{
?
}
\description{
Determines infection spread between farms for a given day?
}
