% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_depopulation_output.R
\name{he_initialize_depopulation_output}
\alias{he_initialize_depopulation_output}
\title{Initialize depopulation output}
\usage{
he_initialize_depopulation_output(
  environment,
  filepath,
  file_name = "depopulated_farms.txt"
)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the depopulation output file}

\item{file_name}{name to use for creating and writing the depopulation
output file}
}
\value{
NA
}
\description{
Initialize depopulation output
}
