% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_inf_farm_info.R
\name{he_initialize_inf_farm_info}
\alias{he_initialize_inf_farm_info}
\title{Initialize matrix to store and track infected farm information}
\usage{
he_initialize_inf_farm_info(environment)
}
\arguments{
\item{environment}{simulation environment in which to store the infected
farm info}
}
\value{
NA
}
\description{
Initialize matrix to store and track infected farm information
}
