% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infection.R
\name{he_initialize_infection}
\alias{he_initialize_infection}
\title{Initialize index infection and start tracking infected farms}
\usage{
he_initialize_infection(
  inf_farm_info,
  simulation_env,
  num_index_infected_min,
  num_index_infected_mode,
  num_index_infected_max,
  species_info,
  farm_info,
  index_netpen_ids,
  type_of_contact = "direct"
)
}
\arguments{
\item{inf_farm_info}{data frame of infected farm information}

\item{simulation_env}{environment containing simulation variables}

\item{num_index_infected_min}{a minimum number of animals to be infected}

\item{num_index_infected_mode}{a mode of animals to be infected}

\item{num_index_infected_max}{a maximum number of animals to be infected}

\item{species_info}{data frame of species information}

\item{farm_info}{data frame of farm information}

\item{index_netpen_ids}{netpen ids selected for index infection}

\item{type_of_contact}{type of contact that resulted in the infection}
}
\value{
infected farm info data frame with infection information for
the index farm appended
}
\description{
Initialize index infection and start tracking infected farms
}
