% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infection.R
\name{he_initialize_infection}
\alias{he_initialize_infection}
\title{Initialize index infection and start tracking infected net pens}
\usage{
he_initialize_infection(
  infected_net_pen_info,
  simulation_env,
  n_index_infected_min,
  n_index_infected_mode,
  n_index_infected_max,
  species_info,
  net_pen_info,
  index_net_pen_ids,
  index_infection_stage,
  clinically_infected_prop = NULL
)
}
\arguments{
\item{infected_net_pen_info}{data frame of infected net pen information}

\item{simulation_env}{environment containing simulation variables}

\item{n_index_infected_min}{a minimum number of animals to be infected}

\item{n_index_infected_mode}{a mode of animals to be infected}

\item{n_index_infected_max}{a maximum number of animals to be infected}

\item{species_info}{data frame of species information}

\item{net_pen_info}{data frame of net pen information}

\item{index_net_pen_ids}{net pen ids selected for index infection}

\item{index_infection_stage}{infection stage in which animals infected by the
index infection should start}

\item{clinically_infected_prop}{the proportion of animals which will enter
the clinical stage upon infection, which is the same as 1 minus the
proportion of animals which will enter the subclinical stage upon
infection}
}
\value{
infected net_pen info data frame with infection information for
the index net_pen(s) appended
}
\description{
Initialize index infection and start tracking infected net pens
}
