% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_infection.R
\name{he_initialize_infection}
\alias{he_initialize_infection}
\title{Initialize index infection and start tracking infected farms}
\usage{
he_initialize_infection(
  inf_farm_info,
  simulation_env,
  n_index_infected_min,
  n_index_infected_mode,
  n_index_infected_max,
  species_info,
  farm_info,
  index_netpen_ids,
  index_infection_stage = "subclinical"
)
}
\arguments{
\item{inf_farm_info}{data frame of infected farm information}

\item{simulation_env}{environment containing simulation variables}

\item{n_index_infected_min}{a minimum number of animals to be infected}

\item{n_index_infected_mode}{a mode of animals to be infected}

\item{n_index_infected_max}{a maximum number of animals to be infected}

\item{species_info}{data frame of species information}

\item{farm_info}{data frame of farm information}

\item{index_netpen_ids}{netpen ids selected for index infection}

\item{index_infection_stage}{infection stage in which animals infected by the
index infection should start}
}
\value{
infected farm info data frame with infection information for
the index farm appended
}
\description{
Initialize index infection and start tracking infected farms
}
