% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_model_env.R
\name{he_initialize_model_env}
\alias{he_initialize_model_env}
\title{Initialize model variables}
\usage{
he_initialize_model_env(
  model_env,
  netpen_info_filepath,
  species_info_filepath,
  connectivity_matrix_filepath = NULL
)
}
\arguments{
\item{model_env}{environment in which to store model level variables}

\item{netpen_info_filepath}{filepath to netpen info file}

\item{species_info_filepath}{filepath to species info file}

\item{connectivity_matrix_filepath}{filepath to connectivity matrix file}
}
\value{
NA
}
\description{
Initialize model variables
}
