% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_model_env.R
\name{he_initialize_model_env}
\alias{he_initialize_model_env}
\title{Initialize model variables}
\usage{
he_initialize_model_env(
  model_env,
  net_pen_info_filepath,
  species_info_filepath
)
}
\arguments{
\item{model_env}{environment in which to store model level variables,
this will be modified by the function}

\item{net_pen_info_filepath}{filepath to net pen info file}

\item{species_info_filepath}{filepath to species info file}
}
\value{
NA
}
\description{
Initialize model variables
}
