% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_simulation_env.R
\name{he_initialize_simulation_env}
\alias{he_initialize_simulation_env}
\title{Initialize simulation environment with necessary variables}
\usage{
he_initialize_simulation_env(
  simulation_env,
  species_info,
  output_dir,
  model_run_id,
  infected_netpen_output_file_name,
  case_fatality_prop,
  simulation_n
)
}
\arguments{
\item{simulation_env}{environment in which simulation variables are stored}

\item{species_info}{data frame of species information}

\item{output_dir}{file path where the infected netpen information output should
be stored}

\item{model_run_id}{identifier for this run of the model}

\item{infected_netpen_output_file_name}{name of the file in which the infected
netpen information should be stored}

\item{case_fatality_prop}{proportion of the infected population that will die}

\item{simulation_n}{identifies which simulation of the model run this
this environment belongs to}
}
\value{
NA
}
\description{
Initialize simulation environment with necessary variables
}
