% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_simulation_env.R
\name{he_initialize_simulation_env}
\alias{he_initialize_simulation_env}
\title{Initialize simulation environment}
\usage{
he_initialize_simulation_env(
  environment,
  farm_info_filepath,
  species_info_filepath,
  dist_mat_filepath,
  output_filepath = "/output",
  detailed_survey_output = FALSE,
  detailed_depop_output = FALSE,
  detailed_preemptive_depop_output = FALSE
)
}
\arguments{
\item{environment}{simulation environment}

\item{farm_info_filepath}{filepath to farm info file}

\item{species_info_filepath}{filepath to species info file}

\item{dist_mat_filepath}{filepath to distance matrix file}
}
\value{
NA
}
\description{
Initialize simulation environment
}
