% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_simulation_env.R
\name{he_initialize_simulation_env}
\alias{he_initialize_simulation_env}
\title{Initialize simulation environment with necessary variables}
\usage{
he_initialize_simulation_env(
  simulation_env,
  species_info,
  net_pen_info,
  net_pen_to_net_pen,
  output_dir,
  model_run_id,
  infected_net_pen_output_file_name,
  clinically_infected_prop,
  simulation_n
)
}
\arguments{
\item{simulation_env}{environment in which simulation variables are stored}

\item{species_info}{data frame of species information}

\item{net_pen_info}{data frame of net pen and farm information}

\item{net_pen_to_net_pen}{daily probability for infection between net pens
transmission within a farm}

\item{output_dir}{file path where the infected net pen information output should
be stored}

\item{model_run_id}{identifier for this run of the model}

\item{infected_net_pen_output_file_name}{name of the file in which the infected
net pen information should be stored}

\item{clinically_infected_prop}{the proportion of animals which will enter
the clinical stage upon infection, which is the same as 1 minus the
proportion of animals which will enter the subclinical stage upon
infection}

\item{simulation_n}{identifies which simulation of the model run this
this environment belongs to}
}
\value{
NA
}
\description{
Initialize simulation environment with necessary variables
}
