% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_simulation_env.R
\name{he_initialize_simulation_env}
\alias{he_initialize_simulation_env}
\title{Initialize simulation environment with necessary variables}
\usage{
he_initialize_simulation_env(environment, num_farms, farm_info, species_info)
}
\arguments{
\item{environment}{environment in which simulation variables are stored}

\item{num_farms}{number of farms}

\item{farm_info}{data frame of farm and netpen information}
}
\value{
NA
}
\description{
Initialize simulation environment with necessary variables
}
