% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_simulation_env.R
\name{he_initialize_simulation_env}
\alias{he_initialize_simulation_env}
\title{Initialize simulation environment with necessary variables}
\usage{
he_initialize_simulation_env(
  simulation_env,
  num_netpens,
  farm_info,
  species_info,
  index_netpens
)
}
\arguments{
\item{simulation_env}{environment in which simulation variables are stored}

\item{num_netpens}{number of netpens}

\item{farm_info}{data frame of farm and netpen information}

\item{species_info}{data frame of species information}

\item{index_netpens}{netpen identifiers for netpens which will start
the outbreak}
}
\value{
NA
}
\description{
Initialize simulation environment with necessary variables
}
