% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_initialize_survey_output.R
\name{he_initialize_survey_output}
\alias{he_initialize_survey_output}
\title{Initialize survey output}
\usage{
he_initialize_survey_output(
  environment,
  filepath,
  survey_output_file_name = "surveyed_farms.txt"
)
}
\arguments{
\item{environment}{simulation environment}

\item{filepath}{path in which to store the survey output file}

\item{survey_output_file_name}{name to use for creating and writing the
surveyed farm output file}
}
\value{
NA
}
\description{
Initialize survey output
}
