% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_rpoly2.R
\name{he_rpoly2}
\alias{he_rpoly2}
\title{RPoly2 ?}
\usage{
he_rpoly2(sample_num, probability)
}
\arguments{
\item{sample_num}{number of samples to take}

\item{probability}{numeric probability or vector of probabilities}
}
\value{
number of samples distributed into each bin by probability
}
\description{
RPoly2 ?
}
