% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_update_disease_stage_counts.R
\name{he_update_disease_stage_counts}
\alias{he_update_disease_stage_counts}
\title{Update counts of animals transitioning from one disease stage into another}
\usage{
he_update_disease_stage_counts(
  disease_stage_counts,
  disease_stage_duration_matrices,
  n_newly_infected,
  clinically_infected_prop
)
}
\arguments{
\item{disease_stage_counts}{data frame containing the counts of animals in
each disease stage, where each disease stage is a column and each row is
a netpen}

\item{disease_stage_duration_matrices}{list of matrices containing the
disease stage duration for each netpen}

\item{n_newly_infected}{the number of newly infected animals for each
netpen}

\item{clinically_infected_prop}{the proportion of animals which will enter
the clinical stage upon infection, which is the same as 1 minus the
proportion of animals which will enter the subclinical stage upon
infection}
}
\value{
modified disease stage count columns
}
\description{
Update counts of animals transitioning from one disease stage into another
}
