% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_depopulation_output.R
\name{he_write_depopulation_output}
\alias{he_write_depopulation_output}
\title{Write depopulation output to file}
\usage{
he_write_depopulation_output(environment, output_dir = "", append = FALSE)
}
\arguments{
\item{environment}{simulation environment}

\item{append}{logical indicating whether to overwrite an existing file or
append data}
}
\value{
NA
}
\description{
Write depopulation output to file
}
