% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he_write_infected_netpen_output.R
\name{he_write_inf_netpen_output}
\alias{he_write_inf_netpen_output}
\title{Write infected netpen output to file}
\usage{
he_write_inf_netpen_output(
  inf_netpen_matrix,
  inf_netpen_output_file_name = "infected_netpens.txt",
  output_dir = "",
  append = FALSE
)
}
\arguments{
\item{inf_netpen_matrix}{matrix containing infected netpen data, initialized
and updated over the course of a simulation run}

\item{inf_netpen_output_file_name}{string to name the infected netpen output
file}

\item{output_dir}{string to indicate the directory in which the output file
should be written}

\item{append}{logical indicating whether to overwrite an existing file or
append data}
}
\value{
NA
}
\description{
Write infected netpen output to file
}
