% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_farm_info_file.R
\name{read_farm_info_file}
\alias{read_farm_info_file}
\title{Read in a farm information file}
\usage{
read_farm_info_file(filepath)
}
\arguments{
\item{filepath}{a filepath to read a csv file from}
}
\value{
data.frame of the csv file
}
\description{
Read in a farm information file
}
\examples{
farm_info_filepath <- "data/farm_file_bay_x.csv"
read_farm_info_file(farm_info_filepath)
}
